refactor: Move security related utils under security folder
