+++ createdoc.sh
@@ -5,7 +5,13 @@ parent_path=$( cd "$(dirname "${BASH_SOURCE[0]}")" ; pwd -P )
     cd "$parent_path/lib"
     echo "generating ./doc/alert.md"
     npx jsdoc2md -f ./base/alert.js > ../doc/alert.md
-
+ 
+    echo "generating ./doc/googlesafetynet.md"
+    npx jsdoc2md -f ./security/googlesafetynet.js > ../doc/security/googlesafetynet.md
+    
+    echo "generating ./doc/rootdetection.md"
+    npx jsdoc2md -f ./security/rootdetection.js > ../doc/security/rootdetection.md
+   
     for filename in *.js; do
         if [[ "$filename" == "index.js" ]]; then
             continue
