+++ lib/security/rootdetection.js
@@ -1,40 +1,44 @@
-/*globals SF, requireClass, __SF_RootDetection, array*/
+/*globals requireClass, __SF_RootDetection, array */
 
 /**
  * Smartface RootDetection Module
  * @module rootdetection
  * @type {object}
- * @author Dogan Ekici  <doga.ekici@smartface.io>
- * @author Muhammed Yalcin Kuru  <yalcin.kuru@smartface.io>
- * @copyright Smartface 2018
+ * @author Dogan Ekici <dogan.ekici@smartface.io>
+ * @author Muhammed Yalcin Kuru <yalcin.kuru@smartface.io>
+ * @author Ozcan Ovunc <ozcan.ovunc@smartface.io>
+ * @copyright Smartface 2019
  */
 
 const System = require("sf-core/device/system");
 
 /**
- * RootDetection includes a few established methods to capture whether device is rooted or not. To know if device is rooted, looking for potentially dangerous app packages/paths,
- * system folder accessibility, su binaries & schemas. The dangerous app packages/paths should be kept upto date by commuity.
+ * RootDetection includes a few established methods to capture whether device is 
+ * rooted or not. To know if device is rooted, looking for potentially dangerous 
+ * app packages/paths, system folder accessibility, su binaries & schemas. The 
+ * dangerous app packages/paths should be kept upto date by commuity.
  * 
  * @public
  * @class
  * @example
  * 
- * const RootDetection = require("sf-extension-utils/lib/rootdetection");
+ * const RootDetection = require("sf-extension-utils/lib/security/rootdetection");
  * 
- * if (RootDetection.isRooted())
- * console.log(" The device is found untrusted condition");
- * else
- * console.log(" The device seems safe");
+ * if (RootDetection.isRooted()) {
+ *     console.log("Attention your device is not trusted.");
+ * } else {
+ *     console.log("It seems you can trust your device");
+ * }
  * 
  */
 class RootDetection {
-
 	constructor() {
 		if (System.OS === "iOS") {
 			this.rootDetection = new __SF_RootDetection();
 		}
 		else {
-			const RootDetectionUtil = requireClass("io.smartface.android.utils.RootDetectionUtil");
+			const RootDetectionUtil = requireClass(
+				"io.smartface.android.utils.RootDetectionUtil");
 			this.rootDetection = new RootDetectionUtil();
 		}
 	}
@@ -127,10 +131,10 @@ class RootDetection {
 	 * @instance
 	 * @example
 	 * 
-	 * if (RootDetection.isRooted()){
-	 * 	console.log("Attention your device is not trusted.");
-	 * }else {
-	 * 	console.log("It seems you can trust");
+	 * if (RootDetection.isRooted()) {
+	 *     console.log("Attention your device is not trusted.");
+	 * } else {
+	 *     console.log("It seems you can trust your device");
 	 * }
 	 *
 	 */
@@ -158,7 +162,8 @@ class RootDetection {
 			"/data/",
 			"/dev/"
 		];
-		let isThereSuBinaries = this.rootDetection.checkSuBinaryExistance(array(suBinaryPaths, "java.lang.String"));
+		let isThereSuBinaries = this.rootDetection.checkSuBinaryExistance(
+			array(suBinaryPaths, "java.lang.String"));
 		return isThereSuBinaries;
 	}
 }
